agentserver/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.10
requests>=2.32.3
pydantic-ai==0.0.14
google-generativeai>=0.8.0
python-dotenv==1.0.0